pyproject.toml
[tool.poetry]
name = "certif-task"
version = "0.1.0"
description = ""
authors = ["Norilka <NorilkaRus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
